% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battery.R
\name{bat}
\alias{bat}
\title{battery}
\usage{
bat(var, data)
}
\arguments{
\item{var}{A variable}
}
\value{
A tibble
}
\description{
This is a backend helper function to be used in the greater battery function
}
