% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{check}
\usage{
check(var, data)
}
\arguments{
\item{var}{is a variable}
}
\value{
A row with survey weighted frequency of yes, no, and grid item text.
}
\description{
This is a backend helper function to be used in the greater check all that apply function
}
