% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_battery.R
\name{grid_battery}
\alias{grid_battery}
\title{grid_battery}
\usage{
grid_battery(var, data)
}
\arguments{
\item{var}{A variable of interest from survey design.}
}
\value{
A data.table output of survey proportions for provided variable.
}
\description{
This is a helper function to create grid battery question table.
}
