% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_check.R
\name{grid_check}
\alias{grid_check}
\title{grid_check}
\usage{
grid_check(var, data)
}
\arguments{
\item{var}{A variable of interest from survey design.}
}
\value{
A data.table output of survey proportions for provided variable.
}
\description{
This is a helper function to create grid check all that apply question table.
}
