% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl_check.R
\name{tl_check}
\alias{tl_check}
\title{tl_check}
\usage{
tl_check(vars, data = tl_df)
}
\arguments{
\item{vars}{Vectors of variables from survey data object}

\item{data}{Survey data object}
}
\value{
Returns a tibble of weighted frequencies for given variables.
}
\description{
Creates a table of weight frequencies for check all that apply questions
}
\examples{
tl_check(vars = c("var1_1", "var1_2", "var1_3", "var1_4"), data = df)
}
