% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl_readexcel.R
\name{tl_readexcel}
\alias{tl_readexcel}
\title{tl_readexcel}
\usage{
tl_readexcel(filename, tibble = TRUE)
}
\arguments{
\item{filename}{A reference to an excel workbook}

\item{tibble}{Binary parameter to specify type of file. Defaults to list of tibbles, otherwise produces list of dataframes}
}
\value{
A list of tibbles or dataframes to for reference by trend data
}
\description{
tl_readexcel
}
\examples{
trends <- tl_readexcel("C:\\\\documents\\\\file.xlsx")

}
