% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl_round.R
\name{tl_round}
\alias{tl_round}
\title{tl_round}
\usage{
tl_round(x)
}
\arguments{
\item{x}{is a number}
}
\value{
number rounded to percentage, - for zero, * for below .5%
}
\description{
This function is a helper function that rounds a number to AP-NORC style guidelines
}
\examples{
tl_round(.48273)
tib[-1] <- lapply(tib[-1], "apnorc_round")



}
