% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl_setup.R
\name{tl_setup}
\alias{tl_setup}
\title{tl_setup}
\usage{
tl_setup(data, caseids, weights, dates, APNORC = TRUE)
}
\arguments{
\item{data}{dataframe for making of topline}

\item{caseids}{caseid variable of dataframe in form of string}

\item{weights}{weights variable of dataframe in form of string}

\item{dates}{character string of dates}

\item{APNORC}{specifies whether the topline is for AP-NORC or another department. Defaults TRUE}
}
\value{
produces three objects in the environment: a survey object for functions to call, a character string to pipe field dates, and a dataframe of nsizes for each question to pipe into tables
}
\description{
A function that produces several objects in the parent environment for use of the tl_tib and tl_bat functions
}
\examples{
tl_setup(omnibusdata, caseids = "su_id", weights = "weight_ap", dates = "05/14-17/2019", APNORC = TRUE)
}
