% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl_tib.R
\name{tl_tib}
\alias{tl_tib}
\title{tl_tib}
\usage{
tl_tib(vari, data = tl_df, default = TRUE, top = 0, bot = 0,
  na = FALSE, web = FALSE)
}
\arguments{
\item{vari}{A variable for which you want to produce frequencies}

\item{data}{A survey-object dataset}

\item{default}{Default creates generic nets. Specifying `default = FALSE` will remove all nets. Specifying custom nets will remove default netting.}

\item{top}{The number of columns you want to net from the beginning. Ex: `top = 2`}

\item{bot}{The number of columns you want to net from the end. Ex: `bot = 3`}

\item{na}{`na = TRUE` specifies the question has a "Not applicable" category to be excluded from netting.}

\item{web}{`web = TRUE` specifies there is only one "Skipped" category}
}
\value{
Returns a table of survey weighted frequencies with net categories if specified
}
\description{
This function produces a table with survey weighted frequencies and proper netting.
It creates default net categories that one can toggle on and off, the creation of custom nets, and the specification of survey mode.
This is useful for the creation of toplines in R.
}
\examples{
tl_tib("q3", data = df, default = TRUE)
tl_tib("q2", data = df, top = 3, bot = 2, na = TRUE)

}
