% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl_trend.R
\name{tl_trend}
\alias{tl_trend}
\title{tl_trend}
\usage{
tl_trend(vari, tl_df, trend_df, default = TRUE, res = 3, top = 0,
  bot = 0)
}
\arguments{
\item{vari}{A variable for which you want to produce trend frequencies}

\item{tl_df}{A survey-object dataset}

\item{trend_df}{A list of tibbles produced by the tl_readexcel helper function}

\item{default}{Creates default net categories. Defaults to TRUE}

\item{res}{Number of residual categories, i.e. "skipped," "refused," "Don't know." Defaults to 3.}

\item{top}{Custom top net.}

\item{bot}{Custom bot net.}
}
\value{
Returns a tibble of read in trend data and new trend data
}
\description{
A function for combining data that has already been tabularized with data from a survey dataset
}
